data_to_md-main/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF 处理
PyMuPDF==1.23.8

# 图像处理
Pillow==10.2.0
pyvips==2.2.1

# OpenAI SDK (用于 DeepSeek-OCR)
openai==1.10.0

# HTTP 客户端
httpx==0.26.0

# 日志
python-json-logger==2.0.7

# 环境变量
python-dotenv==1.0.0

